requirements.txt
addict==2.4.0
albumentations==1.3.0
basicsr==1.4.2
cityscapesscripts==2.2.1
clip==1.0
einops==0.6.1
h5py==2.10.0
joblib==1.0.1
kornia==0.6.2
lmdb==1.3.0
loguru==0.6.0
lpips==0.1.4
matplotlib==3.3.4
mc==0.0
numpy==1.22.0
omegaconf==2.1.1
onnx==1.12.0
open_clip_torch==2.16.0
opencv_python==4.5.5.62
opencv_python_headless==4.5.5.64
packaging==20.9
parrots==0.1.7
Pillow==9.5.0
prettytable==3.2.0
pytorch_lightning==1.5.0
PyYAML==6.0
regex==2021.4.4
safetensors==0.3.1
scikit_image==0.18.1
scipy==1.6.2
setuptools==59.5.0
tensorboardX==2.5
tifffile==2020.10.1
timm==0.4.12
torch==1.13.1
torch_tb_profiler==0.4.1
torchmetrics==0.7.0
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
xformers==0.0.16
yacs==0.1.8
yapf==0.31.0
